Rename scripts in electricity_demand_timeseries
