+++ src/egon/data/airflow/dags/pipeline.py
@@ -29,8 +29,8 @@ from egon.data.datasets.gas_prod import CH4Production
 from egon.data.datasets.heat_demand import HeatDemandImport
 from egon.data.datasets.heat_etrago import HeatEtrago
 from egon.data.datasets.heat_supply import HeatSupply
-from egon.data.datasets.electricity_demand_timeseries import hh_demand_profiles
-from egon.data.datasets.electricity_demand_timeseries import hh_demand_buildings
+from egon.data.datasets.electricity_demand_timeseries import hh_profiles
+from egon.data.datasets.electricity_demand_timeseries import hh_buildings
 from egon.data.datasets.industrial_gas_demand import IndustrialGasDemand
 from egon.data.datasets.industrial_sites import MergeIndustrialSites
 from egon.data.datasets.industry import IndustrialDemandCurves
@@ -389,21 +389,21 @@ with airflow.DAG(
     etrago_input_data >> solar_rooftop_etrago
     map_zensus_grid_districts >> solar_rooftop_etrago
 
-    # Household electricity demand Profiles
+    # Household electricity demand profiles
     mv_hh_electricity_load_2035 = PythonOperator(
         task_id="MV-hh-electricity-load-2035",
-        python_callable=hh_demand_profiles.mv_grid_district_HH_electricity_load,
+        python_callable=hh_profiles.mv_grid_district_HH_electricity_load,
         op_args=["eGon2035", 2035, "0.0.0"],
         op_kwargs={"drop_table": True},
     )
 
     mv_hh_electricity_load_2050 = PythonOperator(
         task_id="MV-hh-electricity-load-2050",
-        python_callable=hh_demand_profiles.mv_grid_district_HH_electricity_load,
+        python_callable=hh_profiles.mv_grid_district_HH_electricity_load,
         op_args=["eGon100RE", 2050, "0.0.0"],
     )
 
-    hh_demand_profiles_setup = hh_demand_profiles.setup(
+    hh_demand_profiles_setup = hh_profiles.setup(
         dependencies=[
             vg250_clean_and_prepare,
             zensus_misc_import,
@@ -412,25 +412,25 @@ with airflow.DAG(
             demandregio,
             osm_buildings_streets_preprocessing,
         ],
-        tasks=(hh_demand_profiles.houseprofiles_in_census_cells,
+        tasks=(hh_profiles.houseprofiles_in_census_cells,
                mv_hh_electricity_load_2035,
                mv_hh_electricity_load_2050,
                )
     )
     hh_demand_profiles_setup.insert_into(pipeline)
     householdprofiles_in_cencus_cells = tasks[
-        "electricity_demand_timeseries.hh_demand_profiles.houseprofiles-in-census-cells"
+        "electricity_demand_timeseries.hh_profiles.houseprofiles-in-census-cells"
     ]
     mv_hh_electricity_load_2035 = tasks["MV-hh-electricity-load-2035"]
     mv_hh_electricity_load_2050 = tasks["MV-hh-electricity-load-2050"]
 
-    # Household electricity demand building assignment
-    hh_demand_buildings_setup = hh_demand_buildings.setup(
+    # Household electricity demand buildings
+    hh_demand_buildings_setup = hh_buildings.setup(
         dependencies=[householdprofiles_in_cencus_cells],
     )
 
     hh_demand_buildings_setup.insert_into(pipeline)
-    map_houseprofiles_to_buildings = tasks["electricity_demand_timeseries.hh_demand_buildings.map-houseprofiles-to-buildings"]
+    map_houseprofiles_to_buildings = tasks["electricity_demand_timeseries.hh_buildings.map-houseprofiles-to-buildings"]
 
     # Industry
